pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "preon-systems-cell"
version = "0.1.0"
description = "ATP-centric, deterministic cell simulation engine."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
simulate = "preon_systems_cell.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
